add uninstall button to geode mod

+++ loader/src/loader/ModImpl.cpp
@@ -393,12 +393,20 @@ Result<> Mod::Impl::disable() {
     return Ok();
 }
 
-Result<> Mod::Impl::uninstall() {
+Result<> Mod::Impl::uninstall(bool deleteSaveData) {
     if (m_requestedAction != ModRequestedAction::None) {
         return Err("Mod already has a requested action");
     }
 
-    m_requestedAction = ModRequestedAction::Uninstall;
+    if (this->getID() == "geode.loader") {
+        utils::game::launchLoaderUninstaller(deleteSaveData);
+        utils::game::exit();
+        return Ok();
+    }
+
+    m_requestedAction = deleteSaveData ?
+        ModRequestedAction::UninstallWithSaveData :
+        ModRequestedAction::Uninstall;
 
     std::error_code ec;
     ghc::filesystem::remove(m_metadata.getPath(), ec);
@@ -408,11 +416,21 @@ Result<> Mod::Impl::uninstall() {
         );
     }
 
+    if (deleteSaveData) {
+        ghc::filesystem::remove_all(this->getSaveDir(), ec);
+        if (ec) {
+            return Err(
+                "Unable to delete mod's save directory: " + ec.message()
+            );
+        }
+    }
+
     return Ok();
 }
 
 bool Mod::Impl::isUninstalled() const {
-    return m_requestedAction == ModRequestedAction::Uninstall;
+    return m_requestedAction == ModRequestedAction::Uninstall ||
+        m_requestedAction == ModRequestedAction::UninstallWithSaveData;
 }
 
 ModRequestedAction Mod::Impl::getRequestedAction() const {
